Support Protobuf <3.7.0 lacking headers port_def.inc and port_undef.inc
There are two Protobuf version macros: GOOGLE_PROTOBUF_VERSION and
PROTOBUF_VERSION without the "GOOGLE_" prefix.  Both ancient
and recent only expose macro GOOGLE_PROTOBUF_VERSION with the "GOOGLE_"
prefix but not the other, so we depart from PROTOBUF_VERSION.

Motivation for this change was that headers port_def.inc and
port_undef.inc were only introduced to Protobuf in commit
6bbe197e9c1b6fc38cbdc45e3bf83fa7ced792a3 of release >=3.7.0 while
we want to support Protobuf 3.6.1.3 of Ubuntu 20.04, too.

Signed-off-by: Sebastian Pipping <sebastian@pipping.org>